+++ src/templates/artifacts/SimpleJpaController.groovy
@@ -15,7 +15,7 @@ class $className {
     }
 
     def listAll = {
-        edt {
+        execInsideUIAsync {
             model.${domainClassAsProp}List.clear()
 <%
     fields.collect { field ->
@@ -35,7 +35,7 @@ class $className {
         }
     }
 %>
-        edt {
+        execInsideUIAsync {
             model.${domainClassAsProp}List.addAll(${domainClassAsProp}Result)
             model.searchMessage = app.getMessage("simplejpa.search.all.message")
 <%
@@ -52,9 +52,9 @@ class $className {
 
     def search = {
         if (model.${firstField}Search?.length() > 0) {
-            edt { model.${domainClassAsProp}List.clear() }
+            execInsideUIAsync { model.${domainClassAsProp}List.clear() }
             List result = find${domainClass}By${firstFieldUppercase}(model.${firstField}Search)
-            edt {
+            execInsideUIAsync {
                 model.${domainClassAsProp}List.addAll(result)
                 model.searchMessage = app.getMessage("simplejpa.search.result.message", ['${firstFieldNatural}', model.${firstField}Search])
             }
@@ -82,43 +82,36 @@ class $className {
                 return_failed()
             }
             persist(${domainClassAsProp})
-            edt { model.${domainClassAsProp}List << ${domainClassAsProp} }
+            execInsideUIAsync { model.${domainClassAsProp}List << ${domainClassAsProp} }
         } else {
             // Update operation
-            if (JOptionPane.showConfirmDialog(view.mainPanel, app.getMessage("simplejpa.dialog.update.message"),
-                    app.getMessage("simplejpa.dialog.update.title"), JOptionPane.YES_NO_OPTION, JOptionPane.WARNING_MESSAGE) == JOptionPane.YES_OPTION) {
-                ${domainClass} selected${domainClass} = model.${domainClassAsProp}Selection.selected[0]
+            ${domainClass} selected${domainClass} = model.${domainClassAsProp}Selection.selected[0]
 <%
     out << fields.collect { field ->
         if (field.info=="DOMAIN_CLASS") {
-            return "\t\t\t\tselected${domainClass}.${field.name} = model.${field.name}.selectedItem"
+            return "\t\t\tselected${domainClass}.${field.name} = model.${field.name}.selectedItem"
         } else if (field.type.toString()=="List" && field.info!="UNKNOWN") {
-            return "\t\t\t\tselected${domainClass}.${field.name}.clear()\n" +
-                   "\t\t\t\tselected${domainClass}.${field.name}.addAll(model.${field.name}.selectedValues)"
+            return "\t\t\tselected${domainClass}.${field.name}.clear()\n" +
+                   "\t\t\tselected${domainClass}.${field.name}.addAll(model.${field.name}.selectedValues)"
         } else {
-            return "\t\t\t\tselected${domainClass}.${field.name} = model.${field.name}"
+            return "\t\t\tselected${domainClass}.${field.name} = model.${field.name}"
         }
     }.join("\n")
 %>
-                merge(selected${domainClass})
-            }
+            merge(selected${domainClass})
         }
-        edt { model.clear() }
+        execInsideUIAsync { model.clear() }
     }
 
     def delete = {
         ${domainClass} ${domainClassAsProp} = model.${domainClassAsProp}Selection.selected[0]
 
-        if (JOptionPane.showConfirmDialog(view.mainPanel, app.getMessage("simplejpa.dialog.delete.message"),
-            app.getMessage("simplejpa.dialog.delete.title"), JOptionPane.YES_NO_OPTION, JOptionPane.WARNING_MESSAGE)
-            == JOptionPane.YES_OPTION) {
 <% if (softDelete) {
-        out << "\t\t\t\tsoftDelete${domainClass}(${domainClassAsProp}.id)\n"
+        out << "\t\tsoftDelete${domainClass}(${domainClassAsProp}.id)\n"
    } else {
-        out << "\t\t\t\tdef ${domainClassAsProp}Persist = merge(${domainClassAsProp})\n"
-        out << "\t\t\t\tremove(${domainClassAsProp}Persist)"  } %>
-                edt { model.${domainClassAsProp}List.remove(${domainClassAsProp}) }
-        }
+        out << "\t\tdef ${domainClassAsProp}Persist = merge(${domainClassAsProp})\n"
+        out << "\t\tremove(${domainClassAsProp}Persist)"  } %>
+        execInsideUIAsync { model.${domainClassAsProp}List.remove(${domainClassAsProp}) }
     }
 
 }